Use context to inject kontist client
