+++ console-src/src/App.tsx
@@ -2,16 +2,19 @@ import React from "react";
 import { Router, Route, Switch } from "react-router-dom";
 import { createBrowserHistory } from "history";
 
+import { KontistClientProvider } from "./enhancers/withKontistClient";
 import Dashboard from "./pages/Dashboard";
 
 const history = createBrowserHistory();
 
 const App = () => (
   <Router history={history}>
-    <Switch>
-      <Route path="/" exact component={Dashboard} />
-      <Route path="/dashboard" exact component={Dashboard} />
-    </Switch>
+    <KontistClientProvider>
+      <Switch>
+        <Route path="/" exact component={Dashboard} />
+        <Route path="/dashboard" exact component={Dashboard} />
+      </Switch>
+    </KontistClientProvider>
   </Router>
 );
